Wrap snapcraft builds in a script
The snapcraft go plugin ignores the pinned vendor packages. This causes
problems when the API changes between the lastest release and the master
git branch, which recently happened with the xurls package.

From now on, packages are built with make using the glide-managed
dependencies, then the binary is dumped into a snap package.