Respond to requests right away
If a Telegram bot doesn't respond to a request on its webhook endpoint
before a certain timeout, Telegram will try again. Squirrelbot used to
wait for a successful download before responding, but Telegram doesn't
need to wait for that.

This change spins off the download functions to their own goroutines,
which allows the request handler to return a 200 OK message right away.

Fixes #12